+++ CHANGELOG.md
@@ -5,5 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1.0] Unreleased
+## [0.1.2] Beta release 03
+### Added
+ - Added <mpaa>XXX</mpaa> tag in nfo metadata
+
+### Changed
+ -Changed <rating> tag in nfo metadata to legacy format
+
+### Fixed
+ - Fix R18 series title scrape not working properly
+
+## [0.1.1] Beta release 02 - 11-30-2019
+### Fixed
+ - Fix on try/catch for directory sort
+
+## [0.1.0] - 11-30-2019
+### Added
+ - Initial functionality release
+ - Web searching with -Find
+ - File and directory sort with -Path and -Apply